RavenDB-18503 moving files around
