+++ src/Raven.Server/Documents/Handlers/RachisDatabaseHandler.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents.Handlers.Processors;
+using Raven.Server.Documents.Handlers.Processors.Rachis;
 using Raven.Server.Routing;
 
 namespace Raven.Server.Documents.Handlers
